[REVIEW-39] 🏗️refactor: feature based 기반 폴더 구조 변경
- 시작점인 App에 대한 app 폴더 생성 및 하위에 router 관련 폴더 설정
- 사진, 폰트 같은 정적인 자료를 보관하는 asset 폴더 생성
- 전체에서 공유할 components 폴더
- 기능 기반으로 분리할 feature 폴더 생성
- 공유할 hook들이 들어있는 hooks 폴더 생성
- 재사용 가능한 라이브러리들이 들어있는 lib 폴더 생성
- 전역 상태의 store들을 저장하는 stores 폴더 생성
- 전역에서 사용할 types 폴더 생성
- 각종 유틸리티 함수들을 넣어둘 utils 폴더 생성
